chore: reverted flickering commit
