+++ gatsby-browser.js
@@ -1,4 +1,5 @@
 import React from 'react'
+import { createRoot } from 'react-dom/client'
 import Cookies from 'js-cookie'
 import { isMobile } from 'react-device-detect'
 import { Analytics } from '@deriv/analytics'
@@ -19,6 +20,13 @@ import './static/css/noto-sans-arabic.css'
 import './static/css/ubuntu.css'
 
 const is_browser = typeof window !== 'undefined'
+
+export const replaceHydrateFunction = () => {
+    return (element, container) => {
+        const root = createRoot(container)
+        root.render(element)
+    }
+}
 
 const checkDomain = () => {
     return eval(
