+++ src/features/pages/markets/etf/index.tsx
@@ -1,4 +1,4 @@
-import React, { lazy } from 'react'
+import React from 'react'
 import Layout from 'features/components/templates/layout'
 import Steps from './steps'
 import WhyTrade from './why-trade'
@@ -11,8 +11,7 @@ import SignupPublic from 'features/components/templates/signup/with-banner'
 import Flex from 'features/components/atoms/flex-box'
 import OtherMarketsSlider from 'features/components/molecules/other-markets-slider'
 import MarketNav from 'features/components/templates/navigation/market-nav'
-import SuspenseHelper from 'features/components/atoms/suspense-helper'
-const Footer = lazy(() => import('features/components/templates/footer'))
+import Footer from 'features/components/templates/footer/'
 
 const ETFMarket = () => {
     return (
@@ -36,9 +35,7 @@ const ETFMarket = () => {
             <Steps />
             <OtherMarketsSlider current_market="etfs" />
             <SignupPublic />
-            <SuspenseHelper fallback={<></>}>
-                <Footer />
-            </SuspenseHelper>
+            <Footer />
         </Layout>
     )
 }
