+++ src/pages/markets/stock/index.tsx
@@ -7,13 +7,8 @@ import Layout from 'components/layout/layout'
 import { WithIntl } from 'components/localization'
 import { SEO } from 'components/containers'
 import { TGatsbyHead } from 'features/types'
-import { usePageLoaded } from 'components/hooks/use-page-loaded'
 
 const StocksAndIndicesPage = () => {
-    const [is_mounted] = usePageLoaded()
-
-    if (!is_mounted) return null
-
     return (
         <Layout type="noNav">
             <DerivedFXHero
