+++ src/pages/partners/affiliate-ib/index.tsx
@@ -1,12 +1,16 @@
 import React from 'react'
+import styled from 'styled-components'
 import Loadable from '@loadable/component'
+import Hero from './_hero'
 import { WhyUsType } from './_why-us'
 import { faq_schema } from './_faq-schema'
-import AffiliateIbHero from './_affiliate-ib-hero'
-import Numbers from './_numbers'
+import { Header, Text } from 'components/elements'
+import { LinkButton } from 'components/form'
 import Layout from 'components/layout/layout'
-import { SEO, TMetaAttributes } from 'components/containers'
-import { WithIntl } from 'components/localization'
+import { SectionContainer, Container, SEO, TMetaAttributes } from 'components/containers'
+import { Localize, WithIntl } from 'components/localization'
+import { affiliate_signup_url } from 'common/constants'
+import device from 'themes/device'
 import { TString } from 'types/generics'
 import { TGatsbyHead } from 'features/types'
 import useRegion from 'components/hooks/use-region'
@@ -20,11 +24,107 @@ const MoreReason = Loadable(() => import('./_more-reason'))
 const Faq = Loadable(() => import('./_faq'))
 const PartnerCTA = Loadable(() => import('./_partner-cta'))
 
+type ItemsType = {
+    title: string
+    subtitle: TString
+}[]
+
+type StyledLinkButtonProps = {
+    id?: string
+}
+
 const meta_attributes: TMetaAttributes = {
     og_title: '_t_Affiliate and IB programme | Deriv_t_',
     og_description:
         '_t_Join Deriv’s affiliate and IB programmes and get a chance to be a partner with a trusted online trading provider._t_',
 }
+
+const StyledHeader = styled(Header)`
+    white-space: pre-line;
+    width: 100%;
+    max-width: 70rem;
+    margin-bottom: 4rem;
+    @media ${device.tabletS} {
+        font-size: 40px;
+        text-align: start;
+    }
+`
+
+const StyledLinkButton = styled(LinkButton)<StyledLinkButtonProps>`
+    border-radius: 4px;
+    @media ${device.tabletS} {
+        font-size: 14px;
+        padding: 12px 10px;
+        white-space: nowrap;
+    }
+    @media ${device.mobileL} {
+        font-size: 12px;
+    }
+`
+
+const StyledSectionContainer = styled(SectionContainer)`
+    @media ${device.tabletL} {
+        padding-top: 40px;
+        padding-bottom: 40px;
+        padding-right: 16px;
+        padding-left: 16px;
+    }
+`
+
+const SubtitleHeader = styled(Header)`
+    width: 1170px;
+    @media ${device.laptopL} {
+        width: 100%;
+    }
+    @media ${device.tabletL} {
+        font-size: 16px;
+        text-align: justify;
+    }
+`
+const SectionContainerWrapper = styled(SectionContainer)`
+    padding: 4rem;
+    @media ${device.tabletL} {
+        padding: 0;
+    }
+`
+
+const StyledContainer = styled(Container)`
+    width: 100%;
+    direction: ltr;
+    @media ${device.tabletL} {
+        flex-wrap: wrap;
+        margin: 20px 0;
+    }
+`
+const NumberWrapper = styled.div`
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    width: 32.8rem;
+    @media ${device.tabletL} {
+        padding: 15px;
+        width: auto;
+    }
+    @media ${device.mobileS} {
+        padding: 10px;
+        width: auto;
+    }
+`
+
+const StyledText = styled(Text)`
+    padding-top: 8px;
+    @media ${device.tabletL} {
+        padding-top: 0;
+        margin: 12px 0;
+    }
+`
+
+const items: ItemsType = [
+    { title: '60K+', subtitle: '_t_members_t_' },
+    { title: '$47M+', subtitle: '_t_paid out since inception_t_' },
+    { title: '190+', subtitle: '_t_countries_t_' },
+    { title: '1M+', subtitle: '_t_clients_t_' },
+]
 
 type contentType = {
     [T: string]: TString
@@ -79,9 +179,43 @@ const AffiliateIb = () => {
         : why_partner_with_us_items
 
     return (
-        <Layout>
-            <AffiliateIbHero btn_text={content_data.banner_btn} />
-            <Numbers earn_text={content_data.earn_text} />
+        <Layout type="partners" padding_top="10">
+            <Hero>
+                <StyledHeader as="h1" color="white" align="center" lh="1.25" type="display-title">
+                    <Localize translate_text="_t_Partner with a trusted online trading provider_t_" />
+                </StyledHeader>
+                <StyledLinkButton
+                    id="dm-hero-affiliate-signup"
+                    to={affiliate_signup_url}
+                    external
+                    target="_blank"
+                    type="affiliate_sign_up"
+                    secondary
+                >
+                    <Localize translate_text={content_data.banner_btn} />
+                </StyledLinkButton>
+            </Hero>
+            <StyledSectionContainer padding="8rem 0 4rem">
+                <Container direction="column">
+                    <SubtitleHeader as="p" type="sub-section-title" weight="normal" align="center">
+                        <Localize translate_text={content_data.earn_text} />
+                    </SubtitleHeader>
+                    <SectionContainerWrapper>
+                        <StyledContainer justify="space-around">
+                            {items.map(({ title, subtitle }, index) => (
+                                <NumberWrapper key={index}>
+                                    <Header as="p" type="page-title" align="center">
+                                        {title}
+                                    </Header>
+                                    <StyledText align="center">
+                                        <Localize translate_text={subtitle} />
+                                    </StyledText>
+                                </NumberWrapper>
+                            ))}
+                        </StyledContainer>
+                    </SectionContainerWrapper>
+                </Container>
+            </StyledSectionContainer>
             <DerivAffiliateProgramme />
             {!is_eu && (
                 <>
