--- src/pages/regulatory/_type.ts
@@ -1,3 +1,0 @@
-export type RegulatoryProps = {
-    language: string
-}
