+++ src/pages/trade-types/cfds/_start-trading.tsx
@@ -1,14 +1,13 @@
 import React from 'react'
 import styled from 'styled-components'
 import { SmallContainer } from '../components/_style'
-import { cfdItems } from '../content/static/_start-trading'
 import CommonHeaderSection from 'components/elements/common-header-section'
 import Button from 'components/custom/_button'
 import { StepperView } from 'components/elements'
 import useRegion from 'components/hooks/use-region'
 import useHandleSignup from 'components/hooks/use-handle-signup'
 import { useBrowserResize } from 'components/hooks/use-browser-resize'
-import useVisibleContent from 'components/hooks/use-visible-content'
+import { TString } from 'types/generics'
 
 const ButtonContainer = styled.div`
     margin-left: -1rem;
@@ -21,7 +20,25 @@ const StartTrading = () => {
     const { is_eu } = useRegion()
     const [is_mobile] = useBrowserResize()
     const handleSignup = useHandleSignup()
-    const visible_items = useVisibleContent({ content: cfdItems, config: { is_eu } })
+    const CfdItems: { title: TString; subtitle: TString }[] = [
+        {
+            title: '_t_Practise_t_',
+            subtitle: is_eu
+                ? '_t_Open a demo CFDs account and practise with an unlimited amount of virtual funds._t_'
+                : '_t_Open a demo Deriv MT5, Deriv X, or Deriv cTrader account and practise with an unlimited amount of virtual funds._t_',
+        },
+        {
+            title: '_t_Trade_t_',
+            subtitle: is_eu
+                ? '_t_Trade with a real CFDs account. Get access to leverage and trade positions larger than your existing capital._t_'
+                : '_t_Trade with a real Deriv MT5, Deriv X, or Deriv cTrader account and get access to high leverage to trade positions larger than your existing capital._t_',
+        },
+        {
+            title: '_t_Withdraw_t_',
+            subtitle:
+                '_t_Conveniently withdraw your funds through any of our supported withdrawal methods._t_',
+        },
+    ]
 
     return (
         <SmallContainer direction="column" ai="flex-start">
@@ -31,7 +48,7 @@ const StartTrading = () => {
                 margin_title={is_mobile ? '40px 0 30px 0' : '120px 0 50px 0'}
             />
             <StepperContainer>
-                <StepperView items={visible_items} />
+                <StepperView items={CfdItems} />
             </StepperContainer>
             <ButtonContainer>
                 <Button onClick={handleSignup} label="_t_Create free demo account_t_" primary />
